requirements.txt
# Core dependencies
base58
aiohttp
websockets
pandas
flipside-wrapper
python-dotenv
prometheus-client
solana
solders
numpy
scikit-learn
python-json-logger
networkx

# Dev dependencies
pytest
pytest-asyncio>=0.23  
ruff
mypy
black
